moved friend profile loading route users.js -> profile.js, edited route to properly other users profile with new interests format
