+++ project_source_code/src/index.js
@@ -101,7 +101,6 @@ const chatRoutes = require('./routes/chat');
 const photosRoutes = require('./routes/photos');
 const profileRoutes = require('./routes/profile');
 const spotifyRoutes = require('./routes/spotify');
-const userRoutes = require('./routes/users');
 const settingsRoutes = require('./routes/settings');
 const exploreRoutes = require('./routes/explore');
 const learnmoreRoutes = require('./routes/learnMore');
@@ -115,7 +114,6 @@ app.use('/chat', chatRoutes);
 app.use('/photos', photosRoutes);
 app.use('/profile', profileRoutes);
 app.use('/spotify', spotifyRoutes);
-app.use('/users', userRoutes);
 app.use('/settings', settingsRoutes);
 app.use('/swipe', swipeRoutes);
 app.use('/explore', exploreRoutes);
