+++ project_source_code/src/views/pages/profile.hbs
@@ -33,7 +33,7 @@
       <ul class="list-unstyled">
         {{#each friends}}
         <li>
-          <a href="/users/{{this.id}}" class="text-decoration-none text-success">{{this.username}}</a>
+          <a href="/profile/{{this.id}}" class="text-decoration-none text-success">{{this.username}}</a>
         </li>
         {{/each}}
       </ul>
